refactor(scrapers): IMDbイベント取り込みを映画祭非依存に共通化
venice-film-festival.tsの取り込みロジックをimdb-event-award.tsへ移し、
組織名・カテゴリ・回次計算・対象カテゴリ判定を設定で渡す形にした。
収集データのキーもgoldenLionからtargetAwardへ統一。

回次がundefinedの年でセレモニーのupsertが
"No values to set"で落ちる不具合もあわせて修正した。
